// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  VOLUNTEER: 'VOLUNTEER',
  EVENT_MANAGER: 'EVENT_MANAGER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EventStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]


export const RegistrationStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  ATTENDED: 'ATTENDED'
} as const

export type RegistrationStatus = (typeof RegistrationStatus)[keyof typeof RegistrationStatus]


export const PostType = {
  ANNOUNCEMENT: 'ANNOUNCEMENT',
  DISCUSSION: 'DISCUSSION'
} as const

export type PostType = (typeof PostType)[keyof typeof PostType]